+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -8,7 +8,7 @@ public interface FilmStorage {
 
     List<Film> getAllFilms();
 
-    Film getFilmById(Long id);
+    Film getFilmById(Integer id);
 
     Film putFilm(Film task);
